[fix] 修复导出是精度丢失问题

+++ ureport2-core/src/main/java/com/bstek/ureport/export/excel/low/Excel97Producer.java
@@ -161,7 +161,7 @@ public class Excel97Producer {
 				        			cell.setCellType(CellType.STRING);
 				        		}else if(obj instanceof Number){
 				        			BigDecimal bigDecimal=Utils.toBigDecimal(obj);
-				        			cell.setCellValue(bigDecimal.floatValue());
+				        			cell.setCellValue(bigDecimal.doubleValue());
 				        			cell.setCellType(CellType.NUMERIC);
 				        		}else if(obj instanceof Boolean){
 				        			cell.setCellValue((Boolean)obj);
